requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.7
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.4.2
numba==0.53.1
numpy==1.20.3
packaging==21.3
pandas==1.2.4
Pillow==8.2.0
pooch==1.3.0
py-spy==0.3.7
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.31
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
torch
torchaudio
torchvision
typing-extensions==4.0.0
urllib3==1.26.4
PyYAML==6.0
tqdm==4.63.0
asteroid-filterbanks==0.4.0